+++ src/mixins/actions.js
@@ -7,36 +7,71 @@ export default {
     }
   },
   methods: {
-    async exportCsv () {
-      const moduleState = this.$store.state[this.moduleName]
-      const params = {
-        filter: `${moduleState.filter}`,
-        from: Math.floor(moduleState.from / 1000),
-        to: Math.floor(moduleState.to / 1000)
-      }
-      this.isFileCsvLoading = true
-      const messages = await this.$store.dispatch(`${this.moduleName}/getMessages`, params)
-      const colsSchema = moduleState.cols
-      const cols = colsSchema.schemas[colsSchema.activeSchema].cols.map(col => col.name)
-      let csv = ''
-      try {
-        csv = parse(messages, { fields: cols })
-      } catch (e) { console.log(e) }
-      if (csv) {
-        const typeOfFile = 'text/csv',
-          link = document.createElement('a'),
-          content = csv,
-          blob = new Blob([content], { type: typeOfFile }),
-          data = URL.createObjectURL(blob),
-          fileName = `${this.item.name || '*NoName*'}[${date.formatDate(moduleState.from, 'DD.MM.YYYY HH-mm-ss')}-${date.formatDate(moduleState.to, 'DD.MM.YYYY HH-mm-ss')}].csv`
-        link.setAttribute('href', data)
-        link.setAttribute('download', fileName)
-        link.style.display = 'none'
-        document.body.appendChild(link)
-        link.click()
-        document.body.removeChild(link)
-      }
-      this.isFileCsvLoading = false
+    async exportCsv (params, { from, to }) {
+      this.$q.dialog(
+        {
+          title: 'Export CSV',
+          cancel: {
+            color: 'white',
+            flat: true
+          },
+          ok: {
+            color: 'white',
+            flat: true
+          },
+          persistent: true,
+          dark: true,
+          options: {
+            type: 'toggle',
+            model: [],
+            items: [
+              { label: 'Format timestamps', value: 'needFormat', color: 'primary' }
+            ]
+          }
+        }
+      )
+        .onOk((data) => {
+          const moduleState = this.$store.state[this.moduleName]
+          this.isFileCsvLoading = true
+          this.$store.dispatch(`${this.moduleName}/getMessages`, params)
+            .then(messages => {
+              const colsSchema = moduleState.cols
+              const cols = colsSchema.schemas[colsSchema.activeSchema].cols.filter(col => !col.__dest).map(col => col.name)
+              if (data.includes('needFormat')) {
+                messages.forEach(message => {
+                  Object.keys(message).forEach(name => {
+                    if (name.match(/timestamp$/) || name === 'begin' || name === 'end') {
+                      const time = Math.floor(message[name] * 1000)
+                      message[name] = date.formatDate(time, 'DD/MM/YYYY HH:mm:ss')
+                    }
+                  })
+                })
+              }
+              let csv = ''
+              try {
+                csv = parse(messages, { fields: cols })
+              } catch (e) { console.log(e) }
+              if (csv) {
+                const typeOfFile = 'text/csv',
+                  link = document.createElement('a'),
+                  content = csv,
+                  blob = new Blob([content], { type: typeOfFile }),
+                  data = URL.createObjectURL(blob),
+                  fileName = `${this.item.name || '*NoName*'}[${date.formatDate(from, 'DD.MM.YYYY HH-mm-ss')}-${date.formatDate(to, 'DD.MM.YYYY HH-mm-ss')}].csv`
+                link.setAttribute('href', data)
+                link.setAttribute('download', fileName)
+                link.style.display = 'none'
+                document.body.appendChild(link)
+                link.click()
+                document.body.removeChild(link)
+              }
+              this.isFileCsvLoading = false
+            })
+            .catch(e => {
+              console.log(e)
+              this.isFileCsvLoading = false
+            })
+        })
     }
   }
 }